Rewritten Pack.c, eliminating all global variables
Pack.c should now be thread-safe for unpacking and packing,
as its functions pass around an explicitly allocated state.
changes should eventually go back into the RTS code (but we
need to see the performance first).

The PAP case has not been rewritten yet, it should be revised
to pack the bitmap instead of tagging the packed stack elements.