CI: Run "Run failed tests"-step only if "Run tests"-step failed
The old behavior was to run the "Run failed tests"-step in case of failure of _any_ previous step. But this is unnecessary if a step other than the "Run tests"-step failed, and also confusing to look at when viewing a failed CI job (since the output of the "Run failed tests"-step is not relevant unless it was the "Run tests"-step that failed).